fix issue where cell image is incorrect
Because the images for QBItemCell instances are loaded asynchronously,
the imageView.image could be set after the cell has already been
dequeued for re-use. This would cause the thumbnail to be incorrect
and the image chosen by tapping it would be a different image.

By making the code wait until the image is retrieved, we ensure that the
cell's imageView.image has the correct image.